Create utility filter that returns new collection
