+++ troposphere/static/js/utilities/filterCollection.js
@@ -1,0 +1,19 @@
+import backbone from 'backbone';
+import moment from 'moment';
+import ImageVersionCollection from 'collections/ImageVersionCollection';
+
+const filterEndDate = function(list) {
+    let newList = list.filter(
+        (version) => {
+            let dateNow = moment(new Date()).format();
+            let endDate = version.get('end_date')
+            if (!endDate) { return true }
+            if (endDate.format() === "Invalid date") { return true }
+            if (endDate.isAfter(dateNow)) { return true }
+            return false
+        }
+    );
+    return new list.constructor(newList);
+}
+
+export { filterEndDate }
